requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
Django==4.2.4
django-cors-headers==4.2.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
dotenv-python==0.0.1
drf-jwt==1.19.2
drf-yasg==1.21.7
dynaconf==3.2.1
hvac==1.1.1
idna==3.4
inflection==0.5.1
knox==0.1.14
loguru==0.7.0
mysqlclient==2.2.0
packaging==23.1
pycparser==2.21
pyhcl==0.4.4
PyJWT==1.7.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
shortuuid==1.0.11
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.4
validators==0.21.2
